refactor: address review on the 2FA take-over changes
- Split the passkey→code switch into two single-action steps (reveal the other
  options, then choose the authenticator/code method). act() performs one action
  per call, so a combined 'click More options, then Use authenticator app' could
  stop after the first click and never reach the code field.
- Trust an explicit method classification when routing the take-over panel: an
  'other' verification (device approval, CAPTCHA, link) now uses the generic
  finish panel instead of code-entry guidance, even when the outcome is
  needs_2fa. Falls back to the failure code only when no method was detected.
- Extract the page classifiers (classifyLoginOutcome, classifyTwoFactorMethod,
  safeOriginAndPath) into a focused browser-login-classifier module, bringing the
  login file back under the 300-line limit.

Tests moved/updated alongside; 186 browserbase tests green, typecheck clean.
