+++ apps/api/src/browserbase/browser-login-classifier.ts
@@ -1,0 +1,131 @@
+import { z } from 'zod';
+
+type Stagehand = import('@browserbasehq/stagehand').Stagehand;
+
+/**
+ * Origin + path only — drops the query, fragment, and userinfo so secrets that
+ * land in an auth redirect (OAuth `code`/`state`, tokens) are never forwarded to
+ * the model. Returns '' for an unparseable URL.
+ */
+export function safeOriginAndPath(rawUrl: string): string {
+  if (!rawUrl) return '';
+  try {
+    const url = new URL(rawUrl);
+    return `${url.origin}${url.pathname}`;
+  } catch {
+    return '';
+  }
+}
+
+export type SignInOutcome =
+  | 'logged_in'
+  | 'invalid_credentials'
+  | 'needs_2fa'
+  | 'challenge'
+  | 'unknown';
+
+/**
+ * Reads the current page after a sign-in attempt and classifies the outcome, so
+ * the connect flow can tell the user what happened and route them correctly.
+ * Never throws — an unreadable page degrades to 'unknown'.
+ */
+export async function classifyLoginOutcome(
+  stagehand: Stagehand,
+): Promise<SignInOutcome> {
+  try {
+    // Give the model where the browser actually is, so it can judge for itself
+    // whether we're on the real app or still on a sign-in / identity-provider
+    // page (it knows hosts like signin.aws.amazon.com or login.microsoftonline.com
+    // without us hardcoding URL patterns). A hint only — falls back to content.
+    let currentUrl = '';
+    try {
+      const pages = stagehand.context?.pages?.() ?? [];
+      const rawUrl = pages[pages.length - 1]?.url() ?? '';
+      // Only the origin + path is needed to judge "is this a sign-in page". Strip
+      // the query, fragment, and userinfo so OAuth codes / tokens / state that
+      // land in an auth redirect are never forwarded to the model.
+      currentUrl = safeOriginAndPath(rawUrl);
+    } catch {
+      // URL unavailable — classify from page content alone.
+    }
+    const { state } = await stagehand.extract(
+      (currentUrl ? `The browser is currently at this URL: ${currentUrl}\n\n` : '') +
+        'Classify this page after a sign-in attempt, using BOTH the page content AND the URL. ' +
+        'Return exactly one value: ' +
+        '"logged_in" — there is clear evidence the user is signed in to the actual application ' +
+        '(a dashboard, an account/avatar menu, or a "Sign out" control) AND the browser is on the ' +
+        'application itself. If the URL is a sign-in, login, SSO, or identity-provider page, or the ' +
+        'page is blank, loading, redirecting, or still shows a sign-in form or a "Sign in" / ' +
+        '"Log in" button, it is NOT logged_in; ' +
+        '"invalid_credentials" — it shows an incorrect username/email/password error; ' +
+        '"needs_2fa" — it asks for a two-factor, one-time, authenticator, or verification code; ' +
+        '"challenge" — it shows a CAPTCHA, a "verify it\'s you", a device-approval, or an email/SMS link step; ' +
+        '"unknown" — none of the above clearly apply.',
+      z.object({
+        state: z.enum([
+          'logged_in',
+          'invalid_credentials',
+          'needs_2fa',
+          'challenge',
+          'unknown',
+        ]),
+      }),
+    );
+    return state;
+  } catch {
+    return 'unknown';
+  }
+}
+
+/**
+ * How a verification step is asking the user to prove identity, so a take-over
+ * UI can give exact guidance:
+ * - `code`         — a one-time / authenticator / SMS / email code can be entered.
+ * - `passkey`      — a passkey / security key is requested, but another method
+ *                    (a code) can be chosen instead.
+ * - `passkey_only` — a passkey / security key is the ONLY option; can't be
+ *                    completed in the automated browser.
+ * - `other`        — a different verification (device approval, CAPTCHA, or an
+ *                    email/SMS link to click).
+ */
+export type TwoFactorMethod = 'code' | 'passkey' | 'passkey_only' | 'other';
+
+/**
+ * Reads a verification page and classifies HOW it wants the user to verify, so
+ * the connect flow can tell them exactly what to do during a take-over. Runs
+ * only when a human take-over is imminent (a small extra cost on that path).
+ * Never throws — an unreadable or unexpected page degrades to 'other'.
+ */
+export async function classifyTwoFactorMethod(
+  stagehand: Stagehand,
+): Promise<TwoFactorMethod> {
+  try {
+    const { method } = await stagehand.extract(
+      'This page is a sign-in verification step. Classify how it is asking the ' +
+        'user to verify, and whether another method can be chosen. Return exactly one:\n' +
+        '"code" — a one-time / authenticator / SMS / email verification CODE can be ' +
+        'entered right now (a code field is visible);\n' +
+        '"passkey" — it is asking for a passkey or security key, AND there is a way to ' +
+        "switch to another method (a 'More options', 'Try another way', or 'use a code' control);\n" +
+        '"passkey_only" — it is asking for a passkey or security key and NO other method ' +
+        'is offered;\n' +
+        '"other" — a different verification (a device approval/notification, a CAPTCHA, ' +
+        'or an email/SMS link to click).',
+      z.object({
+        method: z.enum(['code', 'passkey', 'passkey_only', 'other']),
+      }),
+    );
+    // Guard against a non-conforming value (e.g. a mocked/edge response) so the
+    // return type always holds.
+    if (
+      method === 'code' ||
+      method === 'passkey' ||
+      method === 'passkey_only'
+    ) {
+      return method;
+    }
+    return 'other';
+  } catch {
+    return 'other';
+  }
+}
